+++ bin/fences-slicer
@@ -1,6 +1,6 @@
 #!/usr/bin/env node
 
-var fs = require('fs');
+var fs = require('fs-extra');
 var path = require('path');
 var async = require('async');
 var geojsonSlicer = require('./../index');
@@ -17,12 +17,28 @@ if (nconf.get('config')) {
 if (! (nconf.get('inputDir') && nconf.get('outputDir') && nconf.get('regions')) ) {
   console.error(colors.red('[Invalid usage]:'),
     'Must specify inputDir, outputDir, and regions. If using config file, use --config=<path>');
-  process.exit(1);
+  process.exit(-1);
 }
 
+/**
+ * Execute slicing process. Create output files with any polygons
+ * overlapping the specified regions.
+ */
 (function run() {
+
+  if (!fs.existsSync(nconf.get('inputDir'))) {
+    console.error(colors.red('[Error]:'), 'Input directory does not exist');
+    process.exit(-1);
+  }
+
   var inputFiles = fs.readdirSync(nconf.get('inputDir'));
 
+  /**
+   *
+   * @param {string} inputFile
+   * @param {function} callback
+   * @returns {*}
+   */
   function processFile(inputFile, callback) {
 
     // skip non geojson files
@@ -59,5 +75,6 @@ function getInputFile(dir, file) {
 }
 
 function getOutputFile(dir, file, region) {
-  return dir + region + '--' + file;
+  fs.ensureDirSync(path.join(dir, region));
+  return path.join(dir, region, file);
 }
