+++ src/intersectionFilterStream.js
